Simplification de la structure + récupération directe du callback par le back
