+++ oauth_pda/serializers.py
@@ -1,7 +1,8 @@
 from rest_framework import serializers
-from oauthlib.oauth2 import WebApplicationClient
-
 from django.conf import settings
+
+from oauthlib.oauth2 import WebApplicationClient
+import uuid
 
 
 class AuthorizationLink:
@@ -29,3 +30,35 @@ class AuthorizationLinkSerializer(serializers.Serializer):
     Serializer pour AuthorizationLink
     """
     link = serializers.CharField()
+
+
+class UserInfo:
+
+    def __init__(self, portailResponse):
+        """
+        portailResponse : réponse du portail sur la route des informations
+        de l'utilisateur
+        """
+        # prénom
+        self.firstname = portailResponse['firstname']
+        # nom
+        self.lastname = portailResponse['lastname']
+        # ID
+        self.id = uuid.UUID(portailResponse['id'])
+
+
+AnonymousUserInfo = UserInfo({
+    'firstname': 'Anonyme',
+    'lastname': '',
+    'id': str(uuid.UUID(int=0))
+})
+
+
+class UserInfoSerializer(serializers.Serializer):
+    """
+    Serializer pour renvoyer au front les informations sur l'utilisateur
+    connecté
+    """
+    firstname = serializers.CharField()
+    lastname = serializers.CharField()
+    id = serializers.UUIDField()
