--- oauth_pda/serializers/__init__.py
@@ -1,14 +1,0 @@
-from .OAuthSerializers import AuthorizationLink
-from .OAuthSerializers import AuthorizationLinkSerializer
-
-from .UserInfoSerializers import UserInfo
-from .UserInfoSerializers import UserInfoSerializer
-from .UserInfoSerializers import AnonymousUserInfo
-
-__all__ = [
-    AuthorizationLink,
-    AuthorizationLinkSerializer,
-    UserInfo,
-    AnonymousUserInfo,
-    UserInfoSerializer,
-]
