--- oauth_pda/views/__init__.py
@@ -1,12 +1,0 @@
-from .oauth_views import GetAuthorizationLink
-from .oauth_views import request_oauth_token
-from .oauth_views import user_logout
-
-from .user_views import GetUserInfo
-
-__all__ = [
-    GetAuthorizationLink,
-    request_oauth_token,
-    user_logout,
-    GetUserInfo,
-]
